Remove Debug metadata from copied instruction to prevent metadata leaks and eventually Module verification failure
When copying the instruction onto the Module meant for the GPU, debug metadata attached to an instruction causes all related metadata to be pulled into the Module, including the DICompileUnit, which is not listed in llvm.dbg.cu of the Module. This fails the verification of the Module and generation of the ASM string.

The only Debug metadata of the instruction, the DebugLoc, is set to null by this patch.